Improve topic management (to be continued).

+++ CoMan/Controllers/TopicController.cs
@@ -1,12 +1,8 @@
-﻿using CoMan.Data;
-using CoMan.Models;
+﻿using CoMan.Models;
 using CoMan.Models.AuxiliaryModels;
 using CoMan.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Primitives;
-using Newtonsoft.Json;
 
 namespace CoMan.Controllers
 {
@@ -76,23 +72,40 @@ namespace CoMan.Controllers
         }
 
         // GET: Topic/Edit/{id}
-        [Authorize(Policy = "RequireAdmin")]
+        [Authorize(Policy = "ModifyTopics")]
         public async Task<ActionResult> EditAsync(int id)
         {
-            return View(await _topicService.GetTopicById(id));
+            TopicModel topicToBeUpdated = await _topicService.GetTopicById(id);
+            if (await _topicService.IsUserAllowedToModifyTopic(topicToBeUpdated))
+            {
+                // TODO Add checking if topic can be modified (there are no current cooperations and coop requests)
+                return View(topicToBeUpdated);
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home"); // TODO Access Denied
+            }
         }
 
         // POST: Topic/Edit/{id}
-        [Authorize(Policy = "RequireAdmin")]
+        [Authorize(Policy = "ModifyTopics")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> EditAsync(int id, TopicModel topic)
+        public async Task<ActionResult> EditAsync(int id, TopicModel updatedTopic)
         {
             try
             {
                 TopicModel topicToBeUpdated = await _topicService.GetTopicById(id);
-                await _topicService.UpdateTopic(topicToBeUpdated, topic);
-                return RedirectToAction(nameof(Index));
+                if (await _topicService.IsUserAllowedToModifyTopic(topicToBeUpdated))
+                {
+                    // TODO Add checking if topic can be modified (there are no current cooperations and coop requests)
+                    await _topicService.UpdateTopic(topicToBeUpdated, updatedTopic);
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    return RedirectToAction("Error", "Home");
+                }
             }
             catch
             {
@@ -101,22 +114,41 @@ namespace CoMan.Controllers
         }
 
         // GET: Topic/Delete/{id}
-        [Authorize(Policy = "RequireAdmin")]
+        [Authorize(Policy = "ModifyTopics")]
         public async Task<ActionResult> DeleteAsync(int id)
         {
-            return View(await _topicService.GetTopicById(id));
+            TopicModel topicToBeDeleted = await _topicService.GetTopicById(id);
+            if (await _topicService.IsUserAllowedToModifyTopic(topicToBeDeleted))
+            {
+                // TODO Add checking if topic can be deleted (there are no any cooperations and coop requests connected with topic)
+                // (if there are, suggest change status to Archived)
+                return View(topicToBeDeleted);
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home"); // TODO Access Denied
+            }
         }
 
         // POST: Topic/Delete/{id}
-        [Authorize(Policy = "RequireAdmin")]
+        [Authorize(Policy = "ModifyTopics")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteAsync(int id, TopicModel topic)
         {
             try
             {
-                await _topicService.DeleteTopic(topic);
-                return RedirectToAction(nameof(Index));
+                if (await _topicService.IsUserAllowedToModifyTopic(topic))
+                {
+                    // TODO Add checking if topic can be deleted (there are no any cooperations and coop requests connected with topic)
+                    // (if there are, suggest change status to Archived)
+                    await _topicService.DeleteTopic(topic);
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    return RedirectToAction("Error", "Home"); // TODO Access Denied
+                }
             }
             catch
             {
